+++ php/src-generated/TestRunFinished.php
@@ -44,6 +44,7 @@ final class TestRunFinished implements JsonSerializable
          * Timestamp when the TestRun is finished
          */
         public readonly Timestamp $timestamp = new Timestamp(),
+        public readonly ?string $testRunStartedId = null,
     ) {
     }
 
@@ -57,11 +58,13 @@ final class TestRunFinished implements JsonSerializable
         self::ensureMessage($arr);
         self::ensureSuccess($arr);
         self::ensureTimestamp($arr);
+        self::ensureTestRunStartedId($arr);
 
         return new self(
             isset($arr['message']) ? (string) $arr['message'] : null,
             (bool) $arr['success'],
             Timestamp::fromArray($arr['timestamp']),
+            isset($arr['testRunStartedId']) ? (string) $arr['testRunStartedId'] : null,
         );
     }
 
@@ -98,6 +101,16 @@ final class TestRunFinished implements JsonSerializable
         }
         if (array_key_exists('timestamp', $arr) && !is_array($arr['timestamp'])) {
             throw new SchemaViolationException('Property \'timestamp\' was not array');
+        }
+    }
+
+    /**
+     * @psalm-assert array{testRunStartedId?: string|int|bool} $arr
+     */
+    private static function ensureTestRunStartedId(array $arr): void
+    {
+        if (array_key_exists('testRunStartedId', $arr) && is_array($arr['testRunStartedId'])) {
+            throw new SchemaViolationException('Property \'testRunStartedId\' was array');
         }
     }
 }