+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- BREAKING CHANGE: Add `id` property to `TestRunStarted`, reference in `TestCase` and `TestRunFinished` ([#102](https://github.com/cucumber/messages/pull/102))
+- Add new `GlobalHookStarted` and `GlobalHookFinished` messages ([#102](https://github.com/cucumber/messages/pull/102))
 
 ## [20.0.0] - 2022-11-14
 ### Changed